Update installer.sh to also use pnpm (see below)
Admittedly this is out of date (sorta), since I haven't documented the now ancient change that requires MongoDB (or any enmap-provider that has persistence). This script is run after_install in the the CI via. Travis, so I want it too to also use pnpm for the performance/consistency. I'll get around to documenting the emap/db stuff someday™ (when I'm bugged about it by someone, basically)